Initial Commit
I use LibreOffice Exclusively and as such, the ODT file is the official Template for this archive. However, the way that I use the Template is to fill in my book and save the resulting ODT. I then save another copy as DOCX. I verify that everything within the DOCx is correct and then I upload that to KDP. There is no reason that you cannot simply use the DOCX template with Microsoft Word to complete your eBook. I offer complete Kindle formatting services. Contact me for rates should you need such a service.